修复移除片姐时出现的一个bug；

+++ src/main/java/io/github/hzhilong/bilibili/backup/api/bean/UserCard.java
@@ -132,7 +132,7 @@ public class UserCard {
             private Integer expire;
             private String imageEnhance;
             private String imageEnhanceFrame;
-            private Integer nPid;
+            private Long nPid;
         }
 
         @NoArgsConstructor
